% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnywhere.R
\name{addNamespace}
\alias{addNamespace}
\title{addNamespace}
\usage{
addNamespace(x)
}
\arguments{
\item{x}{a \link[utils]{getAnywhere} object}
}
\value{
Function \link{addNamespace} returns a \link[base]{character} scalar or \link[base]{vector}.
}
\description{
addNamespace
}
\examples{
addNamespace(getAnywhere('rlnorm'))
addNamespace(getAnywhere('format_perc'))
}
\keyword{internal}
