% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkpp.R
\name{batch_rmarkpp}
\alias{batch_rmarkpp}
\title{\link{batch_rmarkpp}}
\usage{
batch_rmarkpp(n, ..., win = owin(xrange = c(-1, 1), yrange = c(-1, 1)))
}
\arguments{
\item{n}{\link[base]{integer} \link[base]{vector},
numbers of \link[spatstat.geom]{ppp} objects to generate for each set of parameters}

\item{...}{see examples, for now}

\item{win}{\link[spatstat.geom]{owin} window, see function \link[spatstat.random]{rMatClust}, etc.}
}
\description{
\link{batch_rmarkpp}
}
