% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppp.R
\name{.rppp}
\alias{.rppp}
\title{Generate Marked Point Pattern}
\usage{
.rppp(
  n = 1L,
  ...,
  win = owin(xrange = c(-1, 1), yrange = c(-1, 1)),
  element1 = TRUE
)
}
\arguments{
\item{n}{\link[base]{integer} scalar,
number of \link[spatstat.geom]{ppp.object}s to generate.
Default \code{1L}.}

\item{...}{two or more named \link[base]{list}s.
The first \link[base]{list} specifies the parameters to
generate the \eqn{x}- and \eqn{y}-\link[spatstat.geom]{coords}.
The second to last \link[base]{list}s specify the parameters to
generate one or more \link[spatstat.geom]{marks}}

\item{win}{\link[spatstat.geom]{owin} window}

\item{element1}{\link[base]{logical} scalar, whether to return
a \link[spatstat.geom]{ppp.object},
instead of a \link[base]{length}-\code{1L} \link[spatstat.geom]{solist},
when \code{n==1L}. Default \code{TRUE}}
}
\value{
Function \link{.rppp} returns a \link[spatstat.geom]{ppp.object} if \code{(n==1L)&element1},
otherwise returns a \link[base]{length}-\code{n} \link[spatstat.geom]{solist}
(which also has \link[base]{class} \code{'ppplist'}).
}
\description{
To generate \link[spatstat.geom]{ppp.object}(s) with
one or more \link[spatstat.geom]{marks}.
}
\examples{
.rppp(rMatClust = list(kappa = c(10,5), scale = c(.15,.06), mu = c(8,4)))

plot(r1 <- .rppp(
 rMatClust = list(kappa = c(10,5), scale = c(.15,.06), mu = c(8,4)), 
 rlnorm = list(meanlog = c(3,5), sdlog = c(.4,.2)),
 rnbinom = list(size = c(4, 6), prob = c(.3, .1))
))

plot(r2 <- .rppp(
 rpoispp = list(lambda = c(3,6)),
 rlnorm = list(meanlog = c(3,5), sdlog = c(.4,.2)),
 rnbinom = list(size = c(4, 6), prob = c(.3, .1))
))

plot(spatstat.geom::superimpose(r1, r2))

}
