% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppp.R
\name{grouped_rppp}
\alias{grouped_rppp}
\title{Batch Process of Function \link{.rppp}}
\usage{
grouped_rppp(n, ..., win = owin(xrange = c(-1, 1), yrange = c(-1, 1)))
}
\arguments{
\item{n}{\link[base]{integer} \link[base]{vector},
numbers of \link[spatstat.geom]{ppp.object}s to generate for each set of parameters}

\item{...}{see examples, for now}

\item{win}{\link[spatstat.geom]{owin} window}
}
\value{
Function \link{grouped_rppp} returns a \link[spatstat.geom]{hyperframe}.
}
\description{
Batch Process of Function \link{.rppp}
}
