% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns.R
\name{ns_fun_fromImports}
\alias{ns_fun_fromImports}
\title{Attach \link[base:getNamespace]{Namespace} and \code{:} to \emph{Imported} Function Name}
\usage{
ns_fun_fromImports(fun, pkg)
}
\arguments{
\item{fun}{\link[base]{character} scalar or \link[base]{vector}}

\item{pkg}{\link[base]{character} scalar}
}
\description{
Attach \link[base:getNamespace]{Namespace} and \code{:} to \emph{Imported} Function Name
}
\examples{
# ns_fun_fromImports(fun = c('col_green', 'check_rhub'), pkg = 'adv.tzh')
# fine with devtools::check() on \pkg{adv.tzh}
# but not okay if copied into other packages :)
}
\keyword{internal}
