% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns.R
\name{ns_name}
\alias{ns_name}
\title{Attach \link[base:getNamespace]{Namespace} and \code{:} to Function Name}
\usage{
ns_name(x)
}
\arguments{
\item{x}{a \link[utils]{getAnywhere} object}
}
\value{
Function \link{ns_name} returns a \link[base]{character} scalar or \link[base]{vector}.
}
\description{
Attach \link[base:getNamespace]{Namespace} and \code{:} to Function Name
}
\examples{
ns_name(getAnywhere('rlnorm')) # exported
ns_name(getAnywhere('format_perc')) # not exported
}
\keyword{internal}
