% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarks.ppp.R
\name{rfactor}
\alias{rfactor}
\title{Generate Random \link[base]{factor}}
\usage{
rfactor(n, prob, levels = as.character(seq_along(prob)))
}
\arguments{
\item{n}{\link[base]{integer} scalar}

\item{prob}{\link[base]{numeric} \link[base]{vector}, see function \link[base]{sample.int}}

\item{levels}{\link[base]{character} \link[base]{vector}, see function \link[base]{factor}}
}
\description{
..
}
\examples{
rfactor(n = 100L, prob = c(4,2,3))
rfactor(n = 100L, prob = c(4,2,3), levels = letters[1:3])
}
\keyword{internal}
