% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarks.ppp.R
\name{rmarks_ppp_}
\alias{rmarks_ppp_}
\title{Generate Random Marks Generation Function}
\usage{
rmarks_ppp_(f)
}
\arguments{
\item{f}{\link[base]{function} of random number generation,
e.g., \link[stats]{rlnorm}, \link[stats]{rnbinom}, etc.}
}
\value{
Function \link{rmarks_ppp_} returns a \strong{\link[base]{function}}.
}
\description{
Generate Random Marks Generation Function
}
\examples{
rmarks_ppp_(rlnorm)
}
\keyword{internal}
