% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm_.R
\name{rmvnorm_}
\alias{rmvnorm_}
\title{Batch Process of Random Multivariate Normal Generation}
\usage{
rmvnorm_(n, mu, Sigma, ...)
}
\arguments{
\item{n}{\link[base]{integer} scalar}

\item{mu}{\link[base]{list} or \link[base]{data.frame}.
Each element must be a \link[base]{numeric} scalar or \link[base]{vector}}

\item{Sigma}{\link[base]{list} or \link[base]{data.frame}.
Each element must be \link[base]{numeric}, either
a scalar, a \link[base]{vector} or a \link[base]{matrix}.}

\item{...}{additional parameters, currently not in use}
}
\description{
Batch Process of Random Multivariate Normal Generation
}
\examples{
rmvnorm_(n = 5L, mu = list(a = c(0, 3), b = c(3, 7)), Sigma = list(a = .5, b = 1.2))
rmvnorm_(n = 5L, mu = list(a = 0, b = c(3, 7)), Sigma = list(a = .5, b = 1.2))
}
\keyword{internal}
