% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_by.ppp.R
\name{sort_by.ppp}
\alias{sort_by.ppp}
\title{\link{sort_by.ppp}}
\usage{
\method{sort_by}{ppp}(x, y = c("x", "y"), ...)
}
\arguments{
\item{x}{a \link[spatstat.geom]{ppp}}

\item{y}{\link[base]{character} scalar, either \code{'x'} or \code{'y'}}

\item{...}{parameters of function \link[base]{order}}
}
\description{
\link{sort_by.ppp}
}
\keyword{internal}
