% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarks.R
\name{rmarkpp}
\alias{rmarkpp}
\title{Generates Marked Point Pattern}
\usage{
rmarkpp(..., win = owin(xrange = c(-1, 1), yrange = c(-1, 1)))
}
\arguments{
\item{...}{see examples, for now}

\item{win}{\link[spatstat.geom]{owin} window, see function \link[spatstat.random]{rMatClust}, etc.}
}
\value{
Function \link{rmarkpp} returns a \link[spatstat.geom]{ppp} object.
}
\description{
..
}
\examples{
r1 = rmarkpp(
 rMatClust = list(kappa=c(10,5), scale=c(.15,.06), mu=c(8,4)), 
 rlnorm = list(meanlog=c(3,5), sdlog=c(.4,.2))
); plot(r1)

r2 = rmarkpp(
 rpoispp = list(lambda=c(3,6)),
 rlnorm = list(meanlog=c(3,5), sdlog=c(.4,.2))
); plot(r2)

plot(spatstat.geom::superimpose(r1, r2))

}
