% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarks.ppp.R
\name{rmarks_ppp}
\alias{rmarks_ppp}
\alias{rlnorm.ppp}
\alias{rnbinom.ppp}
\title{Generate Random \code{marks} for \link[spatstat.geom]{ppp} Object}
\usage{
rlnorm.ppp(x, meanlog = 0, sdlog = 1)

rnbinom.ppp(x, size, prob, mu)
}
\arguments{
\item{x}{a \link[spatstat.geom]{ppp} object}

\item{meanlog, sdlog}{see function \link[stats]{rlnorm}}

\item{size, prob, mu}{see function \link[stats]{rnbinom}}
}
\value{
All functions return a \link[spatstat.geom]{ppp} object.
}
\description{
Generate random \code{marks} for \link[spatstat.geom]{ppp} object.
}
\details{
Function \link{rlnorm.ppp} generates random log-normal \code{marks} on a \link[spatstat.geom]{ppp} object.

Function \link{rnbinom.ppp} generates random negative-binomial \code{marks} on a \link[spatstat.geom]{ppp} object.
}
