% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm_.R
\name{rmvnorm_}
\alias{rmvnorm_}
\title{Batch Process of Random Multivariate Normal Generation}
\usage{
rmvnorm_(n, mu, Sigma, ...)
}
\arguments{
\item{n}{\link[base]{integer} scalar}

\item{mu}{\link[base]{list}, each element is passed to parameter \code{mu} of function \link{mvrnorm2}}

\item{Sigma}{\link[base]{list}, each element is passed to parameter \code{Sigma} of function \link{mvrnorm2}}

\item{...}{additional parameters of function \link[MASS]{mvrnorm}}
}
\value{
Function \code{\link[=rmvnorm_]{rmvnorm_()}} returns a \link[base]{list} of \link[base]{matrix}es.
}
\description{
Batch process of random multivariate normal generation \link{mvrnorm2}.
}
\examples{
rmvnorm_(n = 5L, mu = list(a = c(0, 3), b = c(3, 7)), Sigma = list(a = .5, b = 1.2))
rmvnorm_(n = 5L, mu = list(a = 0, b = c(3, 7)), Sigma = list(a = .5, b = 1.2))
}
\keyword{internal}
